webapi/webapi-messaging-sysapps-tests/messaging/Messaging_method_reverse.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: messaging_method_reverse</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#smsmanager-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var filter = {
  type: "sms"
};
var messageID = "0123456";
var conversationID = "6543210";

var t1 = async_test("Check if the send method of SmsManager which has no param can not work normally", {timeout: 5000});
var t2 = async_test("Check if the send method of SmsManager which has one param can not work normally", {timeout: 5000});
var t3 = async_test("Check if the segmentInfo method of SmsManager which has no params can not work normally", {timeout: 5000});
var t4 = async_test("Check if the send method of MmsManager which has no params can not work normally", {timeout: 5000});
var t5 = async_test("Check if the fetch method of MmsManager which has no params can not work normally", {timeout: 5000});
var t6 = async_test("Check if the findConversations method which has only one params can not work normally", {timeout: 5000});
var t7 = async_test("Check if the markConversationRead method which has only one params can not work normally", {timeout: 5000});
var t8 = async_test("Check if the deleteConversation method which has no params can not work normally", {timeout: 5000});
var t9 = async_test("Check if the findMessages method which has only one params can not work normally", {timeout: 5000});
var t10 = async_test("Check if the markMessageRead method which has only one params can not work normally", {timeout: 5000});
var t11 = async_test("Check if the getMessage method which has no params can not work normally", {timeout: 5000});
var t12 = async_test("Check if the deleteMessage method which has no params can not work normally", {timeout: 5000});
var t13 = async_test("Check if the send method of SmsManager which has wrong type params can not work normally", {timeout: 5000});
var t14 = async_test("Check if the send method of MmsManager which has wrong type params can not work normally", {timeout: 5000});

t1.step(function () {
  try {
    assert_true(!!message.sms.send, "The define of message");
    message.sms.send().then(function (msg) {
      t1.step(function () {
        assert_unreached("Should not reach here");
      });
      t1.done();
    }, function (error) {
      t1.done();
    });
    setTimeout(function () {
      t1.done();
    }, 4000);
  } catch (error) {
    t1.done();
  }
});

t2.step(function () {
  try {
    assert_true(!!message.sms.send, "The define of message");
    message.sms.send("5556").then(function (msg) {
      t2.step(function () {
        assert_unreached("Should not reach here");
      });
      t2.done();
    }, function (error) {
      t2.done();
    });
    setTimeout(function () {
      t2.done();
    }, 4000);
  } catch (error) {
    t2.done();
  }
});

t3.step(function () {
  try {
    assert_true(!!message.sms.segmentInfo, "The define of message");
    message.sms.segmentInfo().then(function (msg) {
      t3.step(function () {
        assert_unreached("Should not reach here");
      });
      t3.done();
    }, function (error) {
      t3.done();
    });
    setTimeout(function () {
      t3.done();
    }, 4000);
  } catch (error) {
    t3.done();
  }
});

t4.step(function () {
  try {
    assert_true(!!message.mms.send, "The define of message");
    message.mms.send().then(function (msg) {
      t4.step(function () {
        assert_unreached("Should not reach here");
      });
      t4.done();
    }, function (error) {
      t4.done();
    });
    setTimeout(function () {
      t4.done();
    }, 4000);
  } catch (error) {
    t4.done();
  }
});

t5.step(function () {
  try {
    assert_true(!!message.mms.fetch, "The define of message");
    message.mms.fetch().then(function (msg) {
      t5.step(function () {
        assert_unreached("Should not reach here");
      });
      t5.done();
    }, function (error) {
      t5.done();
    });
    setTimeout(function () {
      t5.done();
    }, 4000);
  } catch (error) {
    t5.done();
  }
});

t6.step(function () {
  try {
    assert_true(!!message.findConversations, "The define of message");
    message.findConversations(filter).then(function (msg) {
      t6.step(function () {
        assert_unreached("Should not reach here");
      });
      t6.done();
    }, function (error) {
      t6.done();
    });
    setTimeout(function () {
      t6.done();
    }, 4000);
  } catch (error) {
    t6.done();
  }
});

t7.step(function () {
  try {
    assert_true(!!message.markConversationRead, "The define of message");
    message.markConversationRead(conversationID).then(function (msg) {
      t7.step(function () {
        assert_unreached("Should not reach here");
      });
      t7.done();
    }, function (error) {
      t7.done();
    });
    setTimeout(function () {
      t7.done();
    }, 4000);
  } catch (error) {
    t7.done();
  }
});

t8.step(function () {
  try {
    assert_true(!!message.deleteConversation, "The define of message");
    message.deleteConversation().then(function (msg) {
      t8.step(function () {
        assert_unreached("Should not reach here");
      });
      t8.done();
    }, function (error) {
      t8.done();
    });
    setTimeout(function () {
      t8.done();
    }, 4000);
  } catch (error) {
    t8.done();
  }
});

t9.step(function () {
  try {
    assert_true(!!message.findMessages, "The define of message");
    message.findMessages(filter).then(function (msg) {
      t9.step(function () {
        assert_unreached("Should not find any messages");
      });
      t9.done();
    }, function (error) {
      t9.done();
    });
    setTimeout(function () {
      t9.done();
    }, 4000);
  } catch (error) {
    t9.done();
  }
});

t10.step(function () {
  try {
    assert_true(!!message.markMessageRead, "The define of message");
    message.markMessageRead(messageID).then(function (data) {
      t10.step(function () {
        assert_unreached("Should not reach here");
      });
      t10.done();
    }, function (error) {
      t10.done();
    });
    setTimeout(function () {
      t10.done();
    }, 4000);
  } catch (error) {
    t10.done();
  }
});

t11.step(function () {
  try {
    assert_true(!!message.getMessage, "The define of message");
    message.getMessage().then(function (data) {
      t11.step(function () {
        assert_unreached("Should not reach here");
      });
      t11.done();
    }, function (error) {
      t11.done();
    });
    setTimeout(function () {
      t11.done();
    }, 4000);
  } catch (error) {
    t11.done();
  }
});

t12.step(function () {
  try {
    assert_true(!!message.deleteMessage, "The define of message");
    message.deleteMessage().then(function (data) {
      t12.step(function () {
        assert_unreached("Should not reach here");
      });
      t12.done();
    }, function (error) {
      t12.done();
    });
    setTimeout(function () {
      t12.done();
    }, 4000);
  } catch (error) {
    t12.done();
  }
});

t13.step(function () {
  try {
    assert_true(!!message.sms.send, "The define of message");
    message.sms.send(13813926794, "test").then(function (msg) {
      t13.step(function () {
        assert_unreached("Should not reach here");
      });
      t13.done();
    }, function (error) {
      t13.done();
    });
    setTimeout(function () {
      t13.done();
    }, 4000);
  } catch (error) {
    t13.done();
  }
});

t14.step(function () {
  try {
    assert_true(!!message.mms.send, "The define of message");
    message.mms.send(5566).then(function (msg) {
      t14.step(function () {
        assert_unreached("Should not reach here");
      });
      t14.done();
    }, function (error) {
      t14.done();
    });
    setTimeout(function () {
      t14.done();
    }, 4000);
  } catch (error) {
    t14.done();
  }
});

</script>
